cleanup deleting entries/nodes from the mapp
